requirements.txt
Flask==2.0.2
gunicorn==20.1.0
psycopg2-binary==2.9.1